requirements.txt
asgiref
asttokens
backcall
black
click
colorama
decorator
Django
djangorestframework
executing
ipython
jedi
matplotlib-inline
mypy-extensions
parso
pathspec
pickleshare
platformdirs
prompt-toolkit
pure-eval
pygments
pytz
six
sqlparse
stack-data
tomli
traitlets
typing
tzdata
wcwidth
